+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - 2022-08-03
 
 ### Added
+* Cycle Crossover, a crossover operator for permutations.
 
 ### Changed