+++ CHANGELOG.md
@@ -4,10 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2022-08-03
+## [Unreleased] - 2022-08-04
 
 ### Added
-* Cycle Crossover, a crossover operator for permutations.
+* Crossover operators for permutations:
+  * Cycle Crossover (CX).
+  * Order Crossover (OX).
 
 ### Changed