+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2022-08-11
+## [Unreleased] - 2022-08-12
 
 ### Added
 * Crossover operators for permutations:
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   * Order Crossover (OX)
   * Non-Wrapping Order Crossover (NWOX)
   * Uniform Order-Based Crossover (UOBX)
+  * Order Crossover 2 (OX2)
   * Partially Matched Crossover (PMX)
   * Uniform Partially Matched Crossover (UPMX)
   * Precedence Preservative Crossover (PPX): two-point version