+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2022-08-14
+## [Unreleased] - 2022-08-18
 
 ### Added
 * Crossover operators for permutations:
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   * Precedence Preservative Crossover (PPX): two-point version
   * Precedence Preservative Crossover (PPX): uniform version
   * Edge Recombination
+  * Enhanced Edge Recombination
 
 ### Changed