+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2022-08-30
+## [Unreleased] - 2022-08-31
 
 ### Added
 * Crossover operators for permutations:
@@ -29,7 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 ### Dependencies
-* Bump jpt from 4.0.0 to 4.1.0
+* Bump jpt from 4.0.0 to 4.2.0
 * Bump rho-mu from 2.3.2 to 2.4.1
 
 ### CI/CD